Private

+++ todosite/views.py
@@ -9,7 +9,7 @@ from django.views.generic import TemplateView
 
 def submit(request):
     try:
-        task_list = Task.objects.all()
+        task_list = Task.objects.all().order_by('-date')
         if request.method == "POST":
             formtask = TaskForm(request.POST)
             if formtask.is_valid():
@@ -23,14 +23,34 @@ def submit(request):
         }
     except Task.DoesNotExist:
         raise Http404("Task does not exist")
-
     return render(request, 'pages/todowebsite.html', page)
 
-def remove(request, item_id):
+def deleteall(request):
+    item = Task.objects.all()
+    page = {"item": item}
+    if request.method == "POST":
+        item.delete()
+        return redirect('todosite')
+    return render(request, 'pages/delete.html', page)
+
+def updatetask(request, item_id):
     item = Task.objects.get(id=item_id)
-    item.delete()
-    return redirect('todosite')
+    page = {"item": item}
+    if request.method == "POST":
+        item.title = request.POST.get('title')
+        item.contents = request.POST.get('contents')
+        item.save()
+        return redirect ('todosite')
+    return render(request, 'pages/update.html', page)
 
-class SetUpTask(TemplateView):
-     template_name = "pages/todosetuptask.html"
+def deletetask(request, item_id):
+    item = Task.objects.get(id=item_id)
+    page = {"item": item}
+    if request.method == "POST":
+        item.delete()
+        return redirect('todosite')
+    return render(request, 'pages/delete.html', page)
 
+def site(request):
+    response = redirect('todosite')
+    return response
